backend-python/requirements.txt
fastapi
uvicorn
python-dotenv
PyJWT
agno
duckduckgo-search
groq
yfinance
firecrawl-py
